Reorganize AccessibleAbstractTableView
