+++ src/gui/Src/Accessible/AccessibleAbstractTableView.h
@@ -3,45 +3,9 @@
 #include <QAccessibleWidget>
 #include "../BasicView/AbstractTableView.h"
 
-class AccessibleAbstractTableView;
-
-class AccessibleAbstractTableViewCell : public QAccessibleInterface, QAccessibleTableCellInterface
-{
-    duint row;
-    int column;
-    AccessibleAbstractTableView* mParent;
-public:
-    AccessibleAbstractTableViewCell(AccessibleAbstractTableView* parent, duint row, int column);
-    // QAccessibleInterface
-    QString text(QAccessible::Text t) const override;
-    QColor foregroundColor() const override;
-    QWindow* window() const override;
-    QAccessibleInterface* parent() const override;
-    QAccessibleInterface* child(int index) const override;
-    QAccessibleInterface* childAt(int x, int y) const override;
-    QObject* object() const override;
-    void setText(QAccessible::Text t, const QString & text) override;
-    QRect rect() const override;
-    int indexOfChild(const QAccessibleInterface* child) const override;
-    QAccessible::Role role() const override;
-    QAccessible::State state() const override;
-    int childCount() const override;
-    bool isValid() const override;
-    void* interface_cast(QAccessible::InterfaceType type) override;
-    // QAccessibleTableCellInterface
-    bool isSelected() const override;
-    QList<QAccessibleInterface*> columnHeaderCells() const override;
-    QList<QAccessibleInterface*> rowHeaderCells() const override;
-    int columnIndex() const override;
-    int rowIndex() const override;
-    int columnExtent() const override;
-    int rowExtent() const override;
-    QAccessibleInterface* table() const override;
-};
-
 class AccessibleAbstractTableView : public QAccessibleWidget, public QAccessibleTableInterface
 {
-    mutable std::vector<QAccessible::Id> cellInterfaces;
+    std::vector<QAccessible::Id> cellInterfaces;
     int rows, cols;
     friend class AccessibleAbstractTableViewCell;
     AbstractTableView* m_tableView;
@@ -51,11 +15,10 @@ public:
     // QAccessibleInterface
     int childCount() const override;
     QAccessibleInterface* child(int index) const override;
-    int indexOfChild(const QAccessibleInterface*) const override;
     QAccessibleInterface* QAccessibleInterface::childAt(int x, int y) const override;
+    int indexOfChild(const QAccessibleInterface* child) const override;
     QAccessibleInterface* focusChild() const override;
     bool isValid() const override;
-    QAccessible::Role role() const override;
     QAccessible::State state() const override;
     void* interface_cast(QAccessible::InterfaceType type) override;
     // QAccessibleTableInterface
@@ -80,7 +43,9 @@ public:
     bool unselectColumn(int column) override;
     bool unselectRow(int row) override;
 private:
-    QAccessible::Id & cellArray(int row, int col) const; // throws std::out_of_range exception
+    QAccessible::Id & cellArray(int row, int col); // Get reference of accessible id, throws std::out_of_range exception
+    const QAccessible::Id & cellArray(int row, int col) const;
+    virtual QString getCellContent(int row, int col) const; // Get plain text of a cell
 };
 
 #endif
